Memory allocation macros
Allows the user to replace the functions used to allocate and free memory.

This change is backwards-compatible; if the macros are not defined by the
user, the same functions as before are used.

Only allow to either define all or none of them to make sure the user
does not forget one of them.