Add functionality to find mtl file in-library
Up until now, the callback function written by the client was responsible for finding and loading the material file for the respective obj file. This commit changes this and allows the library to find the material file based on the mtllib listing in the OBJ file.

Add functionality to find mtl file from obj file

Up until now, the callback function written by the client was responsible for finding and loading the material file for the respective obj file. This commit changes this and allows the library to find the material file based on the mtllib listing in the OBJ file.

Add functionality to find mtl file in-library

Up until now, the callback function written by the client was responsible for finding and loading the material file for the respective obj file. This commit changes this and allows the library to find the material file based on the mtllib listing in the OBJ file.