Always grow the hash table if we can't insert an item
If we try to insert an item when the hash table is not completely filled
(i.e. `n` != `capacity`) but the indexes provided by quadratic probing
are all filled, we will end up in an infinite loop. To fix that, always
grow the hash table if we can't insert an item.